decoupled the Twig HttpKernelExtension runtime from the extension
